--- src/components/ingredient-details/ingredient-details.module.css
@@ -1,28 +1,0 @@
-.container {
-	display: flex;
-	flex-direction: column;
-	align-items: center;
-	padding: 0 20px;
-}
-
-.image {
-	width: 480px;
-	height: 240px;
-	object-fit: contain;
-	margin-bottom: 16px;
-}
-
-.nutritionalInfo {
-	display: flex;
-	justify-content: space-between;
-	width: 100%;
-	max-width: 520px;
-}
-
-.nutrient {
-	display: flex;
-	flex-direction: column;
-	align-items: center;
-	min-width: 112px;
-	text-align: center;
-}
